+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The ❗ icon indicates a breaking change.
 
 - Add a type alias: `Result<T> = std::result::Result<T, IntervalError<T>>`.
 
+### Fixed
+
+- Fix Clippy warning `double_neg` caused by passing a negative value to the lower bound of `const_interval!` macro.
+
 ## v0.3.1 - 2020-08-14
 
 ### Fixed